shipper: make shipper aware of block layout
This makes the shipper and remote components more aware of what they are
uploading. The Shipper now attaches a Thanos specific section to the
meta.json file where it attaches the external labels for the block.

The shipper now stages all files in a separate directory, making it less
vulnearable to other processes messing with the currently uploaded
files.t