query: add re-ordering of store input by replica labels
To be able to merge series from different replicas efficiently, we need
to align the sequentially in our series set so we can greedily grab them
until a new series is found.

To fully leverage the streaming API, we'd need to push this logic down
into the store servers, which could attach the (external) replica label
in the way we construct it here. For now this is sufficient without
adding API complexity.