add s3 support to store/sidecar binaries and quickstart.sh
This is how I tested the s3 storage driver, but I'm not married
to anything I did here. I think there might need to be `Close`
function added to the `Bucket` interface to account for needing
to close the `gcsClient`. Then the setup in the `cmd/*` files
could be much more generic.

I also wasn't sure if you all would want the thanos stuff scraped
in the quickstart script, but I found it helpful.