objstore: implement s3 support (#164)
* add minio/minio-go dependency

I added these lines to `Gopkg.toml` and only ran `dep ensure` and it
went a little nuts in `Gopkg.lock`. Let me know if I need to fix that.

* implement objstore.Bucket interface for s3-compatible APIs

Using the minio-go client, implement the wonderful objstore.Bucket
interface.

This was relatively easy, and it appears to work fairly well. I'm
a touch uncertain about the work in GetRange, as the APIs with
gcs didn't quite map as well in that case.

* add s3 support to store/sidecar binaries and quickstart.sh

This is how I tested the s3 storage driver, but I'm not married
to anything I did here. I think there might need to be `Close`
function added to the `Bucket` interface to account for needing
to close the `gcsClient`. Then the setup in the `cmd/*` files
could be much more generic.

I also wasn't sure if you all would want the thanos stuff scraped
in the quickstart script, but I found it helpful.

* addressing PR feedback

- refactor `cmd/thanos/sidecar|store.go` to avoid duplication
- catch a `io.EOF` error in the timingReadCloser
- use `ctx.Done()` in a place it needed to be used
- use `GetObjectWithContext`
- make environment variables better in `scripts/quickstart.sh`
- add note about real S3 APIs in `quickstart.sh`
- refactor s3.Config object to not use pointers
- wrap a couple errors in s3 driver
- rebase and true up `Gopkg.lock`

* sidecar/store: make closeFn more idiomatic

* add s3 capability to rule/compactor

- add s3 flags to rule/compactor
- refactor rule/compactor code to account for multiple backends
- standardize s3 flags across commands
- utilize S3 environment variables in quickstart script

* correct rule policy when no object storage configured

I accidentally made the `rule` node fail if no object storage
was configured, which makes CI fail.