Support binding to any IPv6 interface
Allow Thanos to bind to any interface on both IPv4 and IPv6, for example
by specifying `:7777` as the cluster address rather than `0.0.0.0:7777`.

A lot of the defaults and documentation rely on the IPv4 `0.0.0.0`.
Specifically, the cluster advertise IP address would only be calculated
correctly if `0.0.0.0` were used instead of an empty host name.

This follows idiomatic Go:
https://golang.org/pkg/net/#Dial

> For TCP, UDP and IP networks, if the host is empty or a literal
> unspecified IP address, as in ":80", "0.0.0.0:80" or "[::]:80" for TCP
> and UDP, "", "0.0.0.0" or "::" for IP, the local system is assumed.