Don't allow S3 secret key to be passed as a flag
Passing credentials as command-line flags is not good security practice
as the credentials are visible to any other user on the system.

Using environment variables to pass credentials also has its problems,
but in lieu of using a configuration file, don't allow users to pass the
S3 secret key as a command-line flag.