downsample: fix stale handling, improve chunk count heuristic (#198)
* compact/main: verify compacted and downsampled blocks

This runs the verification procedure whenever we encounter a block that
is not valid. The compactor will block indefinitely and the downsample
run will terminate with an error.

* downsample: fix stale handling, improve chunk count heuristic

* downsample: prevent overlapping batches

* compact: add resolution to metric grouping

* compact: remove per-group metrics

Frequency of compactions is generally too low for histograms over them
to generate sane metrics. Generating metrics for each group may also
result in cardinality blow-up.

Overall the debugging attempted with metrics here is rather a case for
(event) logging since there are just a few compactions per day.

Additionally we re-create groups as of recently which makes
re-registering of the per-group metrics throw panics.

* Address comments