add s3 flags to bucket cmds
I wanted to use these to verify buckets to see why
my compaction was potentially failing, so I added
the s3 bits. Had to refactor a little bit but it
seemed fine.

I also moved the verbose flag definitions into the
s3 implemntation itself, so that doesn't have to be
repeated in every cmd file.