store: Remove mention of hashing in comment
This comment is inaccurate; the store instance is currently run as a
single instance and does no coordination between peers to determine
which subset of data from the object store it should be responsible for
serving.

A single store instance is responsible for all of the data in a single
object store bucket.