reloader: Use `os.Stat` instead of `os.Lstat`
The Thanos reloader ignores any directories inside the rules directory.
If the reloader looks at a symlink to a directory, it is not interpreted as
a directory, but as a symlink and thereby not ignored.

Instead of using `os.Lstat`, use `os.Stat` to follow symlinks before
checking if it is a directory.