Define dependencies for build targets
This patch introduces target dependencies so that requirements are met
for each of the build targets. In practical terms this means there is no
need to explicitly run `make install-tools` or `make deps` as these
targets will be run whevener it is needed. Also, as `make` tracks
modification dates for the real targets, it will not re-run dependent
targets if there were no changes - meaning, these changes should not
impact day-to-day development of thanos.