Define dependencies for build targets (#361)
* Define dependencies for build targets

This patch introduces target dependencies so that requirements are met
for each of the build targets. In practical terms this means there is no
need to explicitly run `make install-tools` or `make deps` as these
targets will be run whevener it is needed. Also, as `make` tracks
modification dates for the real targets, it will not re-run dependent
targets if there were no changes - meaning, these changes should not
impact day-to-day development of thanos.

* Remove install-tools target and amend CI config

The `install-tools` target is no longer necessary as tools are installed
whenever they are first needed.

* Add reasoning for FIRST_GOPATH in Makefile

* Declare test-deps dependency on vendor'd deps

* Add deps target as dependency of the format target

`make format` breaks the code if it's run before `make deps`. This
commit makes that dependency explicit.