Update 'Query Access' section in 'getting_started.md'
I didn't know there was a 'store' API doc, so I'm linking it in this doc (which might as well be the 'landing page' for Thanos, and thereby link to everything you need to know). 

I never understood why you needed to join a cluster so I put some details down, and linked to the relevant section. Also the comments in the code block didn't make much sense either, so I've updated those too.