compaction: add ability to set max compaction level (#578)
On an overloaded scrapers the amount of samples is enormous. Block size
on the 3d compaction level could be more than 30G, hence the next compaction
takes a big effort to download, compact and upload of hundreds Gb (hundreds of billions samples).
This could be essential at least for the undergoing development.