store/cache: add a case for when item > max LRU size
This could potentially spin forever if the size of that item is bigger
than the max size so add this special case to handle it instead of
spinning forever. In this case, the oldest one will not be removed in
this function but later an item might be evicted when we will add an
item to the LRU with c.lru.Add() so it will be OK either way -- just
that we will not spin here forever.