LRU cache ensureFits() tweaks WRT deadlocking (#796)
* store/cache: add a case for when item > max LRU size

This could potentially spin forever if the size of that item is bigger
than the max size so add this special case to handle it instead of
spinning forever. In this case, the oldest one will not be removed in
this function but later an item might be evicted when we will add an
item to the LRU with c.lru.Add() so it will be OK either way -- just
that we will not spin here forever.

* store/cache: return true if items fit in the cache

* ensureFits() now returns true if the passed items will fit in the
cache
* Wire up the results to a new counter metric
thanos_store_index_cache_items_overflowed_total
* Add cache tests to check the edge cases

* store/cache_test: move around names and comments

* store/cache: inline ensureFits() call

No need to create an extra variable to hold a boolean.

* store/cache: use Inc() instead of Add(1)

Inc() is specifically designed for this case so use it instead of
Add(1).