pkg/{store,util}: add safeguards against panics in Series()
Add a new package which will help us safeguard against accidental panics
in the most important Series() call in Thanos Store. Use it in
pkg/store/bucket.go. We already have a top-level handler for panics but
it doesn't work for spawned goroutines so use this one for extra safety.