Skip verification of newly compacted block if allow malformed index
The VerifyIndex() function explicitly states testing of the invariant
ordering, so rather than adding additional parameters to change its
behavior when --debug.accept-malformed-index is set, we skip the
verification step on the newly compacted TSDB block.  This allows
compaction to happen as normal when out of order labels are present in
the index.