compact: accept malformed index (#953)
* Allow malformed index patch

Handle cases where we detect that postings have labels listed
incorrectly due to Prometheus Isuue #5372.  With a command line option
set these specific errors can be ignored as they happen with Prometheus
2.8.0 and lesser versions.

* Fix space in structured logging key

* Fix tests to not ignore out of order label errors in postings

* Skip verification of newly compacted block if allow malformed index

The VerifyIndex() function explicitly states testing of the invariant
ordering, so rather than adding additional parameters to change its
behavior when --debug.accept-malformed-index is set, we skip the
verification step on the newly compacted TSDB block.  This allows
compaction to happen as normal when out of order labels are present in
the index.

* PR feedback for acceptMalformedIndex

* Use fields instead of function parameters
* use the same variable name everywhere

* acceptMalformedIndex: Fix tests reflecting field vs parameters

* Route acceptMalformedIndex option via the Syncer

* accept-malformed-index: PR feedback for comments and error msgs