store/proxy: set {Min,Max}Time accordingly to the nodes
It doesn't make sense to advertise that a Thanos Query node has data
since the beginning of the time and till the end of time if it only has
nodes which have specific time ranges of data.

Change it accordingly by checking each time range of each node and
setting it to the min/max values.