Makefile: fix mac OS build
On one of my mac OS machines the `find . -type f ...` expands to *a lot*
of files, and then all of them are passed as parameters to `goimports`.
Unfortunately, it seems like that hits the default limits of ARG_MAX
i.e. the maximum amount of bytes that can be allocated for the
parameters of a single command line.

Fix this by only passing the directories. `goimports` already descends
to all child directories and checks all of the files. Also, it properly
checks only the Go files as implemented here:
https://github.com/golang/tools/blob/master/cmd/goimports/goimports.go#L64