Docs update
Don't think go depends on git anymore and even if it does the error
message would point in the right direction so no need to mention it
explicitly.

The tutorial is not GCS specific so no need to mention it at the
begining.

The go mod checksum error is already solved in the most recent golang
versions and  has enough google results how to resolve so no need to
mention it here.

Signed-off-by: Krasi Georgiev <8903888+krasi-georgiev@users.noreply.github.com>