Sidecar: Add a limit to how much TSDB data sidecar will query
Usually, long term retention is held in an object storage bucket and it
is most efficient to query that data from the matching Store component.
In cases where Prometheus has more than a few days of configured
retention it can vastly speed up queries to limit the Sidecar queries to
the local Prometheus to just a few days.