pkg/receive: optimize when replication factor is 1
If the replication factor is 1, we can skip the replicate() function
call and directly forward the requests. Replicating to 1 endpoint
still exhibits correct behavior but requires another func call and mem.
Additionally, doing this cleanup means that when a node fails to
store a time series locally, the returned error message is about the
local store rather than about failing to replicate to 1 endpoint.