pkg/receive: ensure hashring chan closes after ctx
Currently, it is possible for the hashring configuration update channel
to be closed before the hashring config watcher has completely stopped.
This means that the config watcher can attempt to write on a closed
chan. This commit fixes that by closing the chan only after the config
watcher will never write again to the chan. This also prevents further
errors of this type by changing the ownership of the chan so that the
goroutine writing to the chan is the one closing it.

This commit also ensures that a closed hashring config chan does not
trigger a new hashring to be sent on the update chan.