compact: add heuristic disk space check
Add a heuristic check to see if there is enough space on the disk to
even perform it. This is needed so that we would avoid downloading
potentially hundreds of gigabytes of data just to see that there is not
enough space. The check is the following: we add up the space that all
of the blocks take in the compaction group and then add another time the
size of the smallest block. It should work well enough.

In addition, improve the metadata file to add how much space a block takes. We do
this by walking through the whole directory recursively and adding up
the amount that files take. This is needed to avoid adding a new method
to our object storage interface.

Tested by breaking out the heuristic check into a separate function and
adding unit tests for it.

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>