cmd,pkg: flush storage when hashring changes (#1480)
* pkg/receive: create flushable storage

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* cmd,pkg: flush storage when hashring changes

In order to allow the hashring of receive nodes to scale at runtime, we
need to temporarily stop the storage, flush it, and upload the blocks.
This commit ensures that whenever a hashring change is detected, the
entire storage is flushed and the shipper is notified so that is uploads
any new blocks. It also ensures that when the receive component starts
up, any in-progress WAL is flushed and uploaded. This ensures that
new data that may belong to a different tenant is not mixed with old
data. Finally, it also ensures that the storage is flushed and uploaded
whenever the process shuts down.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>