pkg/receive: avoid false positive conflict err
This commit ensures that the `replicate` func does not falsely identify
a replication error as a conflict. If the number of conflict errors
during replication is less than the minimum replication threshold, then
replication cannot be said to have failed due to a conflict and the
request should be retried.

Fixes: #1615

xref: #1563
Signed-off-by: Lucas Servén Marín <lserven@gmail.com>