pkg/receive: fix regression + test external API (#1627)
This PR begins by fixing a regression introduced in
19e59ef5369c2134ae3295bd29b86b05b82cb980, where returning a
MultiError threw off the error counting logic in the request
parallelization func, which could cause the wrong HTTP status code to be
returned. In order to protext against these kinds of regressions in the
future, we need to thoroughly test the external API of the receive
component. This commit now fixes the regression as well as adds tests
that verify the HTTP status codes returned by a hashring of receive
nodes and the time series in their DBs when they are in various
configurations and error states.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>