cmd/thanos/bucket: fix static fileserver
This commit fixes a regression introduced in #1702, where the signature
of the bucketui.Register method was changed. The bug was caused because
one of the registered paths relied on the HTTP parameter variable
support from github.com/julienschmidt/httprouter, which the standard
library does not support. The fix is instead to implement the change
recommended in
https://github.com/thanos-io/thanos/pull/1702#discussion_r341269556.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>