store the first raw value of a chunk during downsampling (#1709)
* store the first raw value of a chunk during downsampling

As discussed in #1568, storing only the last raw value
of a chunk will lose a counter reset when:
a) the reset occurs at a chunk boundary, and
b) the last raw value of the earlier chunk is less than
the first aggregated value of the later chunk.

This commit stores the first raw value of a chunk during
the initial raw aggregation, and retains it during
subsequent aggregations. This is similar to the existing
handling for the last raw value of a chunk.

With this change, when counterSeriesIterator iterates over
a chunk boundary, it will see the last raw value of the
earlier chunk, then the first raw value of the later chunk,
and then the first aggregated value of the later chunk. The
first raw value will always be less than or equal to the
first aggregated value, so the only difference in
counterSeriesIterator's output will be the possible detection
of a reset and an extra sample after the chunk boundary.

Fixes: https://github.com/thanos-io/thanos/issues/1568

Signed-off-by: Alfred Landrum <alfred@leakybucket.org>

* changelog for #1709

Signed-off-by: Alfred Landrum <alfred@leakybucket.org>

* adjust existing downsampling tests

Signed-off-by: Alfred Landrum <alfred@leakybucket.org>

* add counter aggregation comments to CounterSeriesIterator

Signed-off-by: Alfred Landrum <alfred@leakybucket.org>