objstore: add initial ObjectSize() implementation
Adds a new method ObjectSize(context.Context, string) method to our
object storage interface which returns the size of the specified object
in uint64 without actually downloading all of it.

This is a pre-requisite to one check that we want to add.

Add a check to our E2E tests to see if it works.

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>