Added special mode where ruler evaluates in past.
This trick works only if the evaluations are within same 2h timestamps. This produces
blocks properly in head for older samples.

This can be actually extended to full backfill one-off job at some point as long as it's done in
same time flow for all rules.

Probably makes sense mostly with vertical compaction enabled rather to join small blocks together for efficiency, but still
will work as it is now.

Signed-off-by: Bartek Plotka <bwplotka@gmail.com>