*: add '+sticky' suffix
Add the ability to specify a '+sticky' suffix via the `--store`
parameter. If it has been specified then the store node's information is
retained between loops of checking and we always consider it healthy.

It is useful in the cases where we run Cortex's query-frontend in front
of Thanos Query and we want to always consider certain store nodes.

We might want to avoid the suffix here and always consider all
statically specified store nodes as sticky but that will be a backward
incompatible change.

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>